Updated permissions for deletion

+++ delete.php
@@ -1,7 +1,12 @@
 <?php include_once ("db.php");
-global $conn;
-$id = mysqli_real_escape_string($conn,$_GET['id']);
-$sql = "DELETE FROM registrations WHERE id='$id'";
-$conn->query($sql);
-$location = "people.php";
-header("Location:" . $location);
+if(!isset($_SESSION['role']) && $_SESSION['role']!="Personnel"){
+    $location = "index.php";
+    header("Location:" . $location);
+} else {
+    global $conn;
+    $id = mysqli_real_escape_string($conn, $_GET['id']);
+    $sql = "DELETE FROM registrations WHERE id='$id'";
+    $conn->query($sql);
+    $location = "people.php";
+    header("Location:" . $location);
+}
